@@ -31,7 +31,6 @@
     "axios": "^1.6.0",
     "axios-mock-adapter": "^1.22.0",
     "classnames": "^2.3.2",
-    "daisyui": "^3.9.2",
     "localforage": "^1.10.0",
     "path-to-regexp": "^6.2.1",
     "react": "^18.2.0",
@@ -45,8 +44,6 @@
     "yup": "^1.2.0"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/typography": "^0.5.9",
     "@testing-library/jest-dom": "^6.1.2",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
@@ -55,16 +52,17 @@
     "@types/jsdom": "^21.1.2",
     "@types/lint-staged": "^13.2.0",
     "@types/node": "^16.18.36",
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.26",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "@vitejs/plugin-react": "^4.1.1",
     "@vitest/coverage-v8": "^0.34.6",
     "@vitest/ui": "^0.34.6",
     "autoprefixer": "^10.4.16",
+    "daisyui": "^3.9.4",
     "eslint": "^8.48.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-typescript": "^3.6.1",
@@ -72,23 +70,24 @@
     "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.1",
+    "eslint-plugin-react-refresh": "^0.4.3",
     "eslint-plugin-tailwindcss": "^3.13.0",
     "husky": "^8.0.3",
     "jsdom": "^22.1.0",
     "lint-staged": "^15.0.2",
     "orval": "^6.17.0",
-    "postcss": "^8.4.31",
     "prettier": "^3.0.3",
     "prettier-plugin-organize-imports": "^3.2.3",
     "tailwind-merge": "^1.13.2",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^4.9.5",
-    "vite": "^4.3.9",
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.0.2",
+    "vite": "^4.5.0",
     "vite-plugin-checker": "^0.6.2",
-    "vite-tsconfig-paths": "^4.2.0",
     "vitest": "^0.34.6"
   },
+  "browserslist": [
+    "defaults"
+  ],
   "lint-staged": {
     "**/*.{ts,tsx,js} !(src/generated/**/*)": "npm run lint:fix",
     "*.{ts,tsx,js,css,md*}": "npm run format"
