@@ -1,32 +1,36 @@
 export const About = () => {
   return (
-    <div className='h-full'>
-      <h1 className='mb-4 text-3xl font-bold'>Welcome to Screen Geometry</h1>
-      <p className='mb-4'>
-        Hey there! We know choosing a monitor can be a real head-scratcher. With a zillion options out there, it&aposs
-        like finding a needle in a haystack. But fret not, I&apos;m here to make your life easier.
-      </p>
-      <p className='mb-4'>
-        So, what&apos;s this all about? Well, I&apos;ve cooked up a handy tool that lets you compare monitor screen
-        sizes and resolutions. No more tech mumbo-jumbo, just plain and simple info to help you decide.
-      </p>
-      <p className='mb-4'>
-        Whether you&apos;re a gamer on the hunt for the perfect screen to level up your gaming experience or a design
-        enthusiast in need of pixel-perfect clarity, I&apos;ve got you covered.
-      </p>
-      <p className='mb-4'>
-        No corporate jargon here! I show you the differences and focus on what matters - finding the ideal monitor for
-        you. Say goodbye to endless scrolling and hello to informed choices.
-      </p>
-      <p className='mb-4'>
-        Thanks for giving me a shot. I&apos;m excited to help you on your monitor-buying journey. Let&apos;s find the
-        perfect screen together!
-      </p>
-      <p className='mb-4'>
-        P.S. Have questions or feedback? I&apos;m all ears. Feel free to reach out and let us know how I can make this
-        experience even better for you.
-      </p>
-      <p className='mb-4 text-right text-sm'>Version: {import.meta.env.PACKAGE_VERSION}</p>
+    <div className='flex h-full flex-col justify-between px-2 md:px-6 xl:px-2'>
+      <div>
+        <h1 className='mb-4 text-3xl font-bold'>Welcome to Screen Geometry</h1>
+        <p className='mb-4'>
+          Hey there! We know choosing a monitor can be a real head-scratcher. With a zillion options out there, it&aposs
+          like finding a needle in a haystack. But fret not, I&apos;m here to make your life easier.
+        </p>
+        <p className='mb-4'>
+          So, what&apos;s this all about? Well, I&apos;ve cooked up a handy tool that lets you compare monitor screen
+          sizes and resolutions. No more tech mumbo-jumbo, just plain and simple info to help you decide.
+        </p>
+        <p className='mb-4'>
+          Whether you&apos;re a gamer on the hunt for the perfect screen to level up your gaming experience or a design
+          enthusiast in need of pixel-perfect clarity, I&apos;ve got you covered.
+        </p>
+        <p className='mb-4'>
+          No corporate jargon here! I show you the differences and focus on what matters - finding the ideal monitor for
+          you. Say goodbye to endless scrolling and hello to informed choices.
+        </p>
+        <p className='mb-4'>
+          Thanks for giving me a shot. I&apos;m excited to help you on your monitor-buying journey. Let&apos;s find the
+          perfect screen together!
+        </p>
+        <p className='mb-4'>
+          P.S. Have questions or feedback? I&apos;m all ears. Feel free to reach out and let us know how I can make this
+          experience even better for you.
+        </p>
+      </div>
+      <div>
+        <p className='mb-4 text-right text-sm'>Version: {import.meta.env.PACKAGE_VERSION}</p>
+      </div>
     </div>
   )
 }
