@@ -1,6 +1,6 @@
 export const Contact = () => {
   return (
-    <div className='h-full'>
+    <div className='h-full px-2 md:px-6 xl:px-2'>
       <h1 className='mb-4 text-3xl font-bold'>How to get hold of me</h1>
     </div>
   )
